Add app form to create and edit app pages. Update styling. Update theme.
