+++ ui/src/App.tsx
@@ -10,13 +10,7 @@ import { currentJhData } from './store';
 import { JhData } from './types/jupyterhub';
 import { getJhData } from './utils/jupyterhub';
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-    },
-  },
-});
+const queryClient = new QueryClient();
 
 export const App = (): React.ReactElement => {
   const [, setJhData] = useRecoilState<JhData>(currentJhData);
