+++ ui/src/components/index.ts
@@ -1,4 +1,5 @@
 export { default as Alert } from './alert/alert';
+export { default as AppForm } from './app-form/app-form';
 export { default as ButtonGroup } from './button-group/button-group';
 export { default as ContextMenu } from './context-menu/context-menu';
 export { default as ErrorMessages } from './error-messages/error-messages';
