+++ ui/src/pages/edit-app/edit-app.tsx
@@ -1,10 +1,32 @@
+import { AppForm } from '@src/components';
+import { APP_BASE_URL } from '@src/utils/constants';
 import React from 'react';
+import { useParams } from 'react-router-dom';
 
 export const EditApp = (): React.ReactElement => {
+  const param = useParams();
   return (
     <div className="container">
-      <div>
-        <h1 className="text-3xl font-bold">Edit app</h1>
+      <div className="row">
+        <h1>Edit app</h1>
+        <p>
+          Edit your app details here. For more information on editing your app,{' '}
+          <span>
+            <a
+              href="https://www.nebari.dev/docs/welcome"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              visit our docs
+            </a>
+          </span>
+        </p>
+      </div>
+      <div className="row">
+        <AppForm
+          id={param.id}
+          onCancel={() => (document.location.href = `${APP_BASE_URL}`)}
+        />
       </div>
     </div>
   );
