+++ ui/src/pages/home/app-card/app-card.tsx
@@ -6,13 +6,13 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import React, { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { ButtonGroup } from '../../../components';
+import AppForm from '../../../components/app-form/app-form';
 import ContextMenu, {
   ContextMenuItem,
 } from '../../../components/context-menu/context-menu';
 import Modal from '../../../components/modal/modal';
 import Tag from '../../../components/tag/tag';
 import { currentNotification } from '../../../store';
-import AppForm from '../app-form/app-form';
 
 interface AppCardProps {
   id: string;
